runall.sh: remove noise from ups.pl command
Current versions of the nut upsc client complain about not using SSL.
I an on localhost and the communication goes over loopback, so I won't
configure SSL.  There seems to be no way to remove the message (lots
of questions on the internet about it), the common solution is
filtering.  So do just that.