diskfree: spawn df(1) for every path
While this spawns more processes than before, it should also stop the
lockups with unreachable NFS mounts and autofs that "df -l -x autofs"
could not prevent.

As only the paths given in the configuration are checked, the "-l"
parameter to df is removed, so you can now check remote paths if you
like (but beware of lockups with sync-mounted NFS shared).